% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3plot.R
\name{draw_heatmap}
\alias{draw_heatmap}
\title{draw a heatmap plot}
\usage{
draw_heatmap(n, group_list)
}
\arguments{
\item{group_list}{A vector with duplicated character or factor}
}
\value{
a heatmap plot according to \code{exp} and grouped by \code{group}.
}
\description{
print a heatmap plot for expression matrix and group by group_list paramter
}
\examples{
#use your example data
exp = matrix(rnorm(60),nrow = 10)
colnames(exp) <- paste0("sample",1:6)
rownames(exp) <- paste0("gene",1:10)
exp[1:4,1:4]
group_list = rep(c("A","B"),each = 3)
draw_heatmap(exp,group_list)
#use iris
n = t(iris[,1:4]);colnames(n) = 1:150
group_list = iris$Species
draw_heatmap(n,group_list)
}
\seealso{
\code{\link{draw_pca}};\code{\link{draw_volcano}}
}
\author{
Xiaojie Sun
}
