% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getCancerStudyFile}
\alias{getCancerStudyFile}
\title{find one cancer study}
\usage{
getCancerStudyFile(
  cancer = NULL,
  study = NULL,
  dataOrigin = NULL,
  uuid = NULL,
  keyword = NULL,
  fileName = NULL,
  analysisSoftware = NULL,
  workflow = NULL
)
}
\arguments{
\item{cancer}{A character(1) canser type}

\item{study}{A character(1) study type}

\item{dataOrigin}{A character(1) database source}
}
\description{
API: /cancer_study/findOne
}
\examples{

getCancerStudyFile("BRAC","transcript","TCGA")

}
