% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{tcgaMutationExpr}
\alias{tcgaMutationExpr}
\title{Correlation between mutation and expression}
\usage{
tcgaMutationExpr(
  cancer,
  gene,
  location = NULL,
  isLocalPath = global_env$isLocalPath
)
}
\arguments{
\item{cancer}{The type of TCGA cancer, such as CHOL}

\item{gene}{gene symbol, such as ARID1A}
}
\value{
a data frame
}
\description{
Correlation between mutation and expression
}
\examples{
plot_data <- tcgaMutationExpr(cancer="CHOL",gene="ARID1A")

boxplot(expr~mutation,data =plot_data,col=c("red","blue"))

ggplot(plot_data,aes(x=mutation,y=expr,color=mutation))+
  stat_boxplot()

}
\author{
YangWang \href{mailto:1749748955@qq.com}{1749748955@qq.com}
}
