% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{tcgaMutationSurvival}
\alias{tcgaMutationSurvival}
\title{The correlation between mutation and survival}
\usage{
tcgaMutationSurvival(
  cancer,
  genes,
  location = NULL,
  isLocalPath = global_env$isLocalPath,
  time = 365
)
}
\description{
The correlation between mutation and survival
}
\examples{
library(survival)
library(survminer)
genes <- c("TP53","ELF3","ARID1A")

df_result <- NULL
for(gene in  genes){
 formula <-as.formula(paste0("Surv(days_to_last_followup,Overall_Survival_Status)~",paste0(gene,collapse = "+")))
  diff <- survdiff(formula,data = survival_df)
  pVal <- 1 -pchisq(diff$chisq,df =1)
  pValue <- signif(pVal,4)
  df_result <- rbind(df_result,data.frame(gene=gene,pValue=pValue))
}

survival_df <- tcgaMutationSurvival("CHOL",genes)
lapply(genes, function(gene) {
  formula <-as.formula(paste0("Surv(days_to_last_followup,Overall_Survival_Status)~",paste0(gene,collapse = "+")))
  fit <- surv_fit(formula, data = survival_df)
  return(list(fit=fit,expr=survival_df,legend=c("Mutation","Wild"),title=gene))
})\%>\%
  lapply(function(fit){
    dev.new()
   res <- ggsurvplot(fit$fit, pval=T, risk.table=F,
                     risk.table.height = 0.3,
                     data = fit$expr,
                     xlab="",
                      pval.size=7,
                      font.x = c(18),
                      font.y = c(18),
                      font.legend=c(18),
                      legend.title = fit$legend,
                      legend.labs = fit$legend,
                      title=fit$title)
    print(res)
  })


}
