% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{initParam}
\alias{initParam}
\title{usage custom url and authorization token}
\usage{
initParam(
  baseUrl = NULL,
  authorization = NULL,
  remote = NULL,
  isLocalPath = NULL
)
}
\description{
usage custom url and authorization token
}
