% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{tcgaMutationExpr}
\alias{tcgaMutationExpr}
\title{Correlation between mutation and expression}
\usage{
tcgaMutationExpr(
  cancer,
  genes,
  location = NULL,
  isLocalPath = global_env$isLocalPath
)
}
\arguments{
\item{cancer}{The type of TCGA cancer, such as CHOL}

\item{gene}{gene symbol, such as ARID1A}
}
\value{
a data frame
}
\description{
Correlation between mutation and expression
}
\examples{
plot_data <- tcgaMutationExpr(cancer="CHOL",gene="ARID1A")
boxplot(expr~mutation,data =plot_data,col=c("red","blue"),main="ARID1A")

plot_data <- tcgaMutationExpr(cancer="CHOL",gene=c("ARID1A","PTEN"))
lapply(plot_data,function(x){
  dev.new()
  boxplot(expr~mutation,data =x$expr,col=c("red","blue"),main=x$gene)
})

}
\author{
YangWang \href{mailto:1749748955@qq.com}{1749748955@qq.com}
}
